moved more of input creation logic to factory module + tidied up imports
