+++ src/domain/collaboration/collaboration-factory/collaboration.factory.module.ts
@@ -1,21 +1,11 @@
 import { Module } from '@nestjs/common';
-import { CalloutModule } from '@domain/collaboration/callout/callout.module';
 import { CollaborationModule } from '../collaboration/collaboration.module';
 import { CollaborationFactoryService } from './collaboration.factory.service';
 import { CalloutFramingModule } from '../callout-framing/callout.framing.module';
-import { CalloutContributionDefaultsModule } from '../callout-contribution-defaults/callout.contribution.defaults.module';
-import { CalloutContributionPolicyModule } from '../callout-contribution-policy/callout.contribution.policy.module';
-import { CalloutContributionModule } from '../callout-contribution/callout.contribution.module';
+import { ProfileModule } from '@domain/common/profile/profile.module';
 
 @Module({
-  imports: [
-    CalloutModule,
-    CollaborationModule,
-    CalloutFramingModule,
-    CalloutContributionModule,
-    CalloutContributionDefaultsModule,
-    CalloutContributionPolicyModule,
-  ],
+  imports: [CollaborationModule, CalloutFramingModule, ProfileModule],
   providers: [CollaborationFactoryService],
   exports: [CollaborationFactoryService],
 })
