+++ src/domain/common/whiteboard/whiteboard.service.ts
@@ -343,17 +343,4 @@ export class WhiteboardService {
 
     return whiteboardContent;
   }
-
-  public createWhiteboardInputFromWhiteboard(
-    whiteboard?: IWhiteboard
-  ): CreateWhiteboardInput | undefined {
-    if (!whiteboard) return undefined;
-    return {
-      profileData: this.profileService.createProfileInputFromProfile(
-        whiteboard.profile
-      ),
-      content: whiteboard.content,
-      nameID: whiteboard.nameID,
-    };
-  }
 }
